+++ examples/publisher/pub_delay/env.json
@@ -1,0 +1,29 @@
+{
+  "queue": {
+    "debugLog": {
+      "on": true,
+      "path": ""
+    },
+    "redis": {
+      "host": "localhost",
+      "port": "6381",
+      "password": "secret",
+      "database": 0,
+      "prefix": "beanq",
+      "maxRetries": 2,
+      "poolsize": 30,
+      "minIdleConnections": 10,
+      "dialTimeout": "5s",
+      "readTimeout": "3s",
+      "writeTimeout": "3s",
+      "poolTimeout": "4s"
+    },
+    "driver": "redis",
+    "poolSize": 10,
+    "jobMaxRetries": 1,
+    "keepJobsInQueue": "3600s",
+    "keepFailedJobsInHistory": "3600s",
+    "keepSuccessJobsInHistory": "3600s",
+    "minWorkers": 10
+  }
+}
